pyproject.toml
[project]
name = "mlops-platform"
version = "0.1.0"
description = "We'll build a great MLOps platform for Machine Learning experimentation!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "catboost>=1.2.10",
    "docker>=7.1.0",
    "dvc>=3.66.1",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "lightgbm>=4.6.0",
    "mlflow>=3.10.0",
    "optuna>=2.10.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "xgboost>=3.2.0",
]


[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
